Move all the online lookup API classes to the `.api` sub-package
