+++ app/src/main/java/org/simple/clinic/patient/onlinelookup/api/RetentionType.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.patient.onlinelookup
+package org.simple.clinic.patient.onlinelookup.api
 
 import com.squareup.moshi.Json
