+++ app/src/main/java/org/simple/clinic/patient/onlinelookup/api/SecondsDuration.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.patient.onlinelookup
+package org.simple.clinic.patient.onlinelookup.api
 
 import com.squareup.moshi.JsonQualifier
 import kotlin.annotation.Retention
